refactor: flatten iron proxy cli config
